feat: support sqlite

+++ src/utils.rs
@@ -66,11 +66,11 @@ pub mod tests {
         }
     }
 
-    pub fn sleep(secs: u64) {
-        std::thread::sleep(std::time::Duration::from_secs(secs));
+    pub fn sleep(secs: i64) {
+        std::thread::sleep(std::time::Duration::from_secs(secs as u64));
     }
 
-    pub async fn sleep_async(secs: u64) {
-        tokio::time::sleep(std::time::Duration::from_secs(secs)).await;
+    pub async fn sleep_async(secs: i64) {
+        tokio::time::sleep(std::time::Duration::from_secs(secs as u64)).await;
     }
 }
